Trait json_in_type::json_value::JSONValue

pub trait JSONValue {
    fn write_json<W: Write>(&self, w: &mut W) -> Result<()>;

    fn to_json_string(&self) -> String { ... }

    fn to_json_buffer(&self) -> Vec<u8> { ... }
}

A trait implemented by types that can be serialized to JSON

Required Methods

fn write_json<W: Write>(&self, w: &mut W) -> Result<()>

Write the object as json to the given writer

Examples

Write a JSON object to a file

use json_in_type::JSONValue;

vec![1, 2, 3].write_json(&mut my_file);

Provided Methods

fn to_json_string(&self) -> String

Returns the object formatted as a json string

fn to_json_buffer(&self) -> Vec<u8>

Returns a buffer containing the bytes of a json representation of the object